Use fs-extra throughout
