+++ config/webpack.config.dev.js
@@ -26,7 +26,6 @@ if (process.env.BUILD_TARGET === 'web') {
   // The trailing $ tells webpack to match 'electron' exactly.
   baseConfig.resolve.alias.process$ = path.join(paths.appSrc, 'utils', 'webShims', 'process.js');
   baseConfig.resolve.alias.electron$ = path.join(paths.appSrc, 'utils', 'webShims', 'electron.js');
-  baseConfig.resolve.alias.fs$ = path.join(paths.appSrc, 'utils', 'webShims', 'fs.js');
   baseConfig.resolve.alias['fs-extra$'] = path.join(paths.appSrc, 'utils', 'webShims', 'fs-extra.js');
   baseConfig.resolve.alias.archiver$ = path.join(paths.appSrc, 'utils', 'webShims', 'archiver.js');
   baseConfig.resolve.alias.decompress$ = path.join(paths.appSrc, 'utils', 'webShims', 'decompress.js');
