+++ src/utils/getAssetData.js
@@ -1,4 +1,4 @@
-import fs from 'fs';
+import fs from 'fs-extra';
 import { memoize } from 'lodash';
 
 const resolver = sourcePath => sourcePath;
